ONT-528 change block num from uint64 to uint32 (#161)

+++ consensus/vbft/msg_pool_test.go
@@ -22,7 +22,7 @@ import "testing"
 
 func TestAddMsg(t *testing.T) {
 	server := constructServer()
-	msgpool := newMsgPool(server, uint64(1))
+	msgpool := newMsgPool(server, 1)
 	block, err := constructBlock()
 	if err != nil {
 		t.Errorf("constructBlock failed :%v", err)
@@ -38,7 +38,7 @@ func TestAddMsg(t *testing.T) {
 
 func TestHasMsg(t *testing.T) {
 	server := constructServer()
-	msgpool := newMsgPool(server, uint64(1))
+	msgpool := newMsgPool(server, 1)
 	block, err := constructBlock()
 	if err != nil {
 		t.Errorf("constructBlock failed :%v", err)
@@ -54,22 +54,22 @@ func TestHasMsg(t *testing.T) {
 
 func TestGetProposalMsgs(t *testing.T) {
 	server := constructServer()
-	msgpool := newMsgPool(server, uint64(1))
-	consensusmsgs := msgpool.GetProposalMsgs(uint64(1))
+	msgpool := newMsgPool(server, 1)
+	consensusmsgs := msgpool.GetProposalMsgs(1)
 	t.Logf("TestGetProposalMsgs: %v", len(consensusmsgs))
 }
 
 func TestGetEndorsementsMsgs(t *testing.T) {
 	server := constructServer()
-	msgpool := newMsgPool(server, uint64(1))
-	consensusmsgs := msgpool.GetEndorsementsMsgs(uint64(1))
+	msgpool := newMsgPool(server, 1)
+	consensusmsgs := msgpool.GetEndorsementsMsgs(1)
 	t.Logf("TestGetEndorsementsMsgs: %v", len(consensusmsgs))
 }
 
 func TestGetCommitMsgs(t *testing.T) {
 	server := constructServer()
-	msgpool := newMsgPool(server, uint64(1))
-	consensusmsgs := msgpool.GetCommitMsgs(uint64(1))
+	msgpool := newMsgPool(server, 1)
+	consensusmsgs := msgpool.GetCommitMsgs(1)
 	t.Logf("TestGetCommitMsgs: %v", len(consensusmsgs))
 }
 
@@ -84,7 +84,7 @@ func TestOnBlockSealed(t *testing.T) {
 	}
 	h, _ := HashMsg(blockproposalmsg)
 	server := constructServer()
-	msgpool := newMsgPool(server, uint64(1))
+	msgpool := newMsgPool(server, 1)
 	t.Logf("TestOnBlockSealed,len:%v", len(msgpool.rounds))
 	if !msgpool.HasMsg(blockproposalmsg, h) {
 		msgpool.AddMsg(blockproposalmsg, h)
